requirements.txt
absl-py==1.4.0
accelerate==1.7.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==4.0.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
av==13.1.0
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.6.2
boto3==1.28.62
botocore==1.31.62
Brotli==1.1.0
build==1.1.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.0.3
cloudpickle==2.2.1
cloup==3.0.7
cmdstanpy==1.2.5
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
config-py==0.1.1
configparser==6.0.0
contourpy==1.0.7
cryptography==40.0.2
cvxpy==1.3.2
cycler==0.11.0
databutton==0.38.3
dataclasses-json==0.5.14
datasets==2.14.5
DateTime==5.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.33.1
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dnspython==2.2.1
duckdb==1.2.2
ecos==2.0.12
ensure==1.0.2
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
Faker==37.1.0
Farama-Notifications==0.0.4
fastapi==0.110.0
fastjsonschema==2.18.1
favicon==0.7.0
feedparser==6.0.10
filelock==3.12.3
filetype==1.2.0
Flask==2.0.2
flask-cors==5.0.1
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.8
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
glcontext==3.0.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.0
grpcio==1.62.1
gspread==5.12.4
gspread-dataframe==4.0.0
gspread-formatting==1.2.1
gspread-pandas==3.3.0
gTTS==2.5.4
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.14.0
h5py==3.9.0
holidays==0.69
htbuilder==0.9.0
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.4
imbalanced-learn==0.12.3
importlib-metadata==6.8.0
importlib_resources==6.3.1
inflate64==0.3.1
ipykernel==6.25.2
ipython==8.14.0
isoduration==20.11.0
isosurfaces==0.1.2
itsdangerous==2.0.1
jedi==0.19.0
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.4.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.4
kubernetes==29.0.0
langchain==0.0.284
langsmith==0.0.44
libclang==16.0.6
libmagic==1.0
lxml==4.9.3
manim==0.19.0
ManimPango==0.6.0
mapbox_earcut==1.0.3
Markdown==3.8
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.9.0.1
mistune==3.0.2
mmh3==4.1.0
moderngl==5.12.0
moderngl-window==3.1.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
multivolumefile==0.2.3
mypy-boto3-s3==1.28.55
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook==7.0.4
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.101.0
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
opt-einsum==3.3.0
orjson==3.9.15
osqp==0.6.3
overrides==7.4.0
packaging==23.0
pandas==2.2.2
pandas-datareader==0.10.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
pillow==10.4.0
platformdirs==3.11.0
plotly==5.15.0
plotly-express==0.4.1
portalocker==2.8.2
posthog==3.5.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
prophet==1.1.6
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.5
pulsar-client==3.4.0
pure-eval==0.2.2
py7zr==0.20.6
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycairo==1.28.0
pycparser==2.21
pycryptodomex==3.19.0
pydantic==1.10.13
pydeck==0.8.0
pydub==0.25.1
pyglet==2.1.6
pyglm==2.8.2
Pygments==2.15.1
PyJWT==2.8.0
pymdown-extensions==10.14.3
pymongo==4.2.0
Pympler==1.0.1
PyMuPDF==1.26.3
pyparsing==3.0.9
PyPika==0.48.9
pyportfolioopt==1.5.5
pyppmd==1.0.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
pyzstd==0.15.9
qdldl==0.1.7.post0
referencing==0.30.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
rouge-score==0.1.2
rpds-py==0.9.2
rsa==4.9
s3transfer==0.7.0
sacrebleu==2.3.1
safetensors==0.5.3
scikit-learn==1.5.1
scipy==1.15.3
screeninfo==0.8.1
scs==3.2.3
seaborn==0.13.2
Send2Trash==1.8.2
sentence-transformers==5.0.0
sgmllib3k==1.0.0
Shimmy==1.2.1
six==1.16.0
skia-pathops==0.8.0.post2
sklearn==0.0.post1
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sql_metadata==2.16.0
SQLAlchemy==2.0.22
sqlparse==0.5.3
srt==3.5.3
st-annotated-text==4.0.2
st-gsheets-connection==0.1.0
st-theme==1.2.3
stable-baselines3==2.1.0
stack-data==0.6.2
stanio==0.5.1
starlette==0.36.3
statsmodels==0.14.0
streamlit==1.38.0
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
svgelements==1.9.6
sympy==1.14.0
ta==0.10.2
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
texttable==1.7.0
threadpoolctl==3.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.52.4
tweepy==4.14.0
typer==0.9.0
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2023.3
tzlocal==4.3.1
unstructured==0.9.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==1.26.15
uuid==1.30
uvicorn==0.18.3
validators==0.34.0
virtualenv==20.26.3
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
Werkzeug==2.0.2
wrapt==1.15.0
xxhash==3.4.1
yarl==1.8.2
yfinance==0.2.23
zipp==3.16.2
zope.interface==6.0
